+++ Algorithms.h
@@ -13,7 +13,7 @@ public:
 	static void reverseSegment(vector<int> & route, int start, int end);
 	static void twoOpt(vector<int> & route, int a, int b);
 	static void swapTwo(vector<int> & route, int a, int b);
-	static bool comparePairBySecond(pair<int, double> & a, pair<int, double> & b);
+	static bool comparePairBySecond(pair<int, long> & a, pair<int, long> & b);
 
 	
 };
